Update DXF handle types to 64-bit throughout codebase
Switch all DXF handle-related fields and methods from 32-bit to 64-bit unsigned integers to support large handle values. This includes entity, owner, material, plot style, image definition, and annotation handles. Updated parsing, assignment, and string conversion logic accordingly. Improves compatibility with modern DXF files using large handle values.

+++ Third Party/libdxfrw/src/EoDxfObjects.h
@@ -40,10 +40,10 @@ class EoDxfObjectEntry {
   void Reset();
 
  public:
-  std::uint32_t m_handle{};  // Group code 5
-  std::uint32_t m_ownerHandle{};  // Group code 330 (soft-pointer to owner dictionary)
-  std::uint32_t m_extensionDictionaryHandle{};  // Group code 360 (hard-owner, extension dictionary)
-  std::vector<std::uint32_t> m_reactorHandles;  // Group code 330 handles within ACAD_REACTORS group
+  std::uint64_t m_handle{};  // Group code 5
+  std::uint64_t m_ownerHandle{};  // Group code 330 (soft-pointer to owner dictionary)
+  std::uint64_t m_extensionDictionaryHandle{};  // Group code 360 (hard-owner, extension dictionary)
+  std::vector<std::uint64_t> m_reactorHandles;  // Group code 330 handles within ACAD_REACTORS group
   std::vector<EoDxfGroupCodeValuesVariant*> m_extensionData;  // Group codes 1000 to 1071
 
  private:
